+++ src/modules/common/src/cli/mod.rs
@@ -1,1 +1,2 @@
+pub mod cli_user_interaction_interface;
 pub mod interface;