+++ src/modules/create/src/service.rs
@@ -1,33 +1,31 @@
-use std::{path::PathBuf, sync::Arc};
-
 use common::core::{
     errors::{Error, Result},
     interfaces::FileSystemInterface,
+    user_interaction_interface::UserInteractionInterface,
 };
-use log::info;
+use std::{path::PathBuf, sync::Arc};
 use templatespecification::core::{
     file_list::FileList,
-    interfaces::Prompt,
     service::TemplateSpecificationService,
     template_specification::{TemplateSpecification, TemplateSpecificationItem},
 };
 
 pub struct CreateTemplateSpecificationService {
     templatespecification_service: Arc<TemplateSpecificationService>,
-    prompt: Arc<dyn Prompt + Send + Sync>,
     file_system: Arc<dyn FileSystemInterface>,
+    user_interaction_interface: Arc<dyn UserInteractionInterface>,
 }
 
 impl CreateTemplateSpecificationService {
     pub fn new(
         templatespecification_service: Arc<TemplateSpecificationService>,
-        prompt: Arc<dyn Prompt + Send + Sync>,
         file_system: Arc<dyn FileSystemInterface>,
+        user_interaction_interface: Arc<dyn UserInteractionInterface>,
     ) -> Self {
         Self {
             templatespecification_service,
-            prompt,
             file_system,
+            user_interaction_interface,
         }
     }
 
@@ -47,18 +45,25 @@ impl CreateTemplateSpecificationService {
         let upload_dir = args.entry_dir;
         match found_creatorly_file {
             Ok(found_template) => {
-                info!("creatorly.yml found, it will only update with the new placeholders");
+                self.user_interaction_interface
+                    .print_success("creatorly.yml found, it will only update with the new placeholders")
+                    .await;
+
                 self.update_template_specification(found_template, files, upload_dir.clone())
                     .await?;
-                info!("🚀 updated creatorly.yml in {}", upload_dir.display().to_string());
             }
             Err(_) => {
-                info!("No creatorly.yml found, it will be created");
+                self.user_interaction_interface
+                    .print("No creatorly.yml found, it will be created")
+                    .await;
+
                 self.create_new_template_specification(files, upload_dir.clone())
                     .await?;
-                info!("🚀 created a creatorly.yml in {}", upload_dir.display().to_string());
             }
         };
+
+        let msg = format!("🚀 updated creatorly.yml in {}", upload_dir.display());
+        self.user_interaction_interface.print(&msg).await;
 
         Ok(())
     }
@@ -90,7 +95,11 @@ impl CreateTemplateSpecificationService {
                 .find(|p| p.template_key == *placeholder);
 
             if found_placeholder.is_none() {
-                let parsed_placeholder = self.prompt.get_default_answer(placeholder)?;
+                let parsed_placeholder = self
+                    .templatespecification_service
+                    .get_default_answer(placeholder)
+                    .await?;
+
                 found_template_specifiaction.placeholders.push(parsed_placeholder);
             }
         }
@@ -146,7 +155,11 @@ impl CreateTemplateSpecificationService {
         // get all answers
         let mut parsed_template_specification_items: Vec<TemplateSpecificationItem> = vec![];
         for placeholder in found_placeholders {
-            let template_specification = self.prompt.get_default_answer(&placeholder)?;
+            let template_specification = self
+                .templatespecification_service
+                .get_default_answer(&placeholder)
+                .await?;
+
             parsed_template_specification_items.push(template_specification);
         }
 
@@ -167,56 +180,57 @@ mod test {
     use std::io::Write;
 
     use super::*;
+    use async_trait::async_trait;
     use common::infrastructure::file_system::FileSystem;
-    use mockall::{mock, predicate::eq};
-    use templatespecification::core::template_specification::{
-        TemplateSpecificationItem, TemplateSpecificationItemType,
+    use mockall::{
+        mock,
+        predicate::{self, eq},
     };
 
     mock! {
-        Prompt {}
+        UserInteractionInterface {}
 
-        impl Prompt for Prompt {
-            fn get_answer(&self, template_specification_item: &mut TemplateSpecificationItem);
+        #[async_trait]
+        impl UserInteractionInterface for UserInteractionInterface {
+            async fn print_success(&self, message: &str);
 
-            fn get_default_answer(&self, placeholder: &str) -> Result<TemplateSpecificationItem>;
+            async fn print_error(&self, message: &str);
+
+            async fn print(&self, message: &str);
+
+            async fn get_input(&self, prompt: &str) -> Result<String>;
         }
     }
 
     #[tokio::test]
     async fn test_create_template_specification() {
         //arrange
-        let mut prompt = MockPrompt::new();
-        prompt
-            .expect_get_default_answer()
-            .with(eq("file_name"))
-            .times(1)
-            .return_const(Ok(TemplateSpecificationItem {
-                template_key: "file_name".to_string(),
-                answer: "file_name".to_string(),
-                item: TemplateSpecificationItemType::SingleChoice("file_name".to_string()),
-            }));
-        prompt
-            .expect_get_default_answer()
-            .with(eq("name"))
-            .times(1)
-            .return_const(Ok(TemplateSpecificationItem {
-                template_key: "name".to_string(),
-                answer: "name".to_string(),
-                item: TemplateSpecificationItemType::SingleChoice("name".to_string()),
-            }));
-        prompt
-            .expect_get_default_answer()
-            .with(eq("description"))
-            .return_const(Ok(TemplateSpecificationItem {
-                template_key: "description".to_string(),
-                answer: "description".to_string(),
-                item: TemplateSpecificationItemType::SingleChoice("description".to_string()),
-            }))
+        let example_dir = get_correct_example_dir().unwrap();
+        let mut mock_user_interaction_interface = MockUserInteractionInterface::new();
+        mock_user_interaction_interface
+            .expect_print()
+            .with(eq("No creatorly.yml found, it will be created"))
+            .return_const(())
+            .times(1);
+        mock_user_interaction_interface
+            .expect_get_input()
+            .with(eq("description: "))
+            .return_const(Ok("description".to_string()));
+        mock_user_interaction_interface
+            .expect_get_input()
+            .with(eq("file_name: "))
+            .return_const(Ok("file_name".to_string()));
+        mock_user_interaction_interface
+            .expect_get_input()
+            .with(eq("name: "))
+            .return_const(Ok("name".to_string()));
+        mock_user_interaction_interface
+            .expect_print()
+            .with(predicate::str::contains("🚀 updated creatorly.yml in /tmp/example"))
+            .return_const(())
             .times(1);
 
-        let example_dir = get_correct_example_dir().unwrap();
-        let sut = get_create_template_service(prompt);
+        let sut = get_create_template_service(Arc::new(mock_user_interaction_interface));
 
         // act
         let args = CreateTemplateArgs {
@@ -243,11 +257,13 @@ mod test {
     async fn test_get_all_creatorly_placeholders() {
         // let mut os_mock = MockFileSystemInterface::new();
         let example_dir = get_correct_example_dir().unwrap();
-
+        let mock_user_interaction_interface = Arc::new(MockUserInteractionInterface::new());
         let service = CreateTemplateSpecificationService::new(
-            Arc::new(TemplateSpecificationService::with_local_file_loader()),
-            Arc::new(MockPrompt::new()),
+            Arc::new(TemplateSpecificationService::with_local_file_loader(
+                mock_user_interaction_interface.clone(),
+            )),
             Arc::new(FileSystem::default()),
+            mock_user_interaction_interface.clone(),
         );
 
         let files = service
@@ -263,11 +279,15 @@ mod test {
         assert!(placeholders.contains(&"description".to_string()));
     }
 
-    fn get_create_template_service(mock_propt: MockPrompt) -> CreateTemplateSpecificationService {
+    fn get_create_template_service(
+        mock_user_interaction_interface: Arc<MockUserInteractionInterface>,
+    ) -> CreateTemplateSpecificationService {
         CreateTemplateSpecificationService::new(
-            Arc::new(TemplateSpecificationService::with_local_file_loader()),
-            Arc::new(mock_propt),
+            Arc::new(TemplateSpecificationService::with_local_file_loader(
+                mock_user_interaction_interface.clone(),
+            )),
             Arc::new(FileSystem::default()),
+            mock_user_interaction_interface.clone(),
         )
     }