+++ src/modules/templatespecification/src/infrastructure/mod.rs
@@ -1,4 +1,3 @@
-pub mod cli_prompt;
 pub mod configuration_loader;
 pub mod folder_loader;
 pub mod liquid_template_renderer;